requirements.txt
streamlit
torch
transformers
pillow
openai
python-dotenv
safetensors